+++ bootique-jdbc/src/main/java/io/bootique/jdbc/JdbcModule.java
@@ -1,7 +1,9 @@
 package io.bootique.jdbc;
 
+import com.google.inject.Binder;
 import com.google.inject.Provides;
 import com.google.inject.Singleton;
+import io.bootique.BQCoreModule;
 import io.bootique.ConfigModule;
 import io.bootique.config.ConfigurationFactory;
 import io.bootique.log.BootLogger;
@@ -10,6 +12,7 @@ import io.bootique.type.TypeRef;
 
 import java.util.HashMap;
 import java.util.Map;
+import java.util.logging.Level;
 
 public class JdbcModule extends ConfigModule {
 
@@ -18,6 +21,21 @@ public class JdbcModule extends ConfigModule {
 
     public JdbcModule(String configPrefix) {
         super(configPrefix);
+    }
+
+    @Override
+    public void configure(Binder binder) {
+
+        // suppress Tomcat PooledConnection logger. It logs some absolutely benign stuff as WARN
+        // per https://github.com/bootique/bootique-jdbc/issues/25
+
+        // it only works partially, namely when SLF intercepts JUL (e.g. under bootique-logback and such).
+
+        // TODO: submit a patch to Tomcat to reduce this log level down...
+
+        BQCoreModule.contributeLogLevels(binder)
+                .addBinding(org.apache.tomcat.jdbc.pool.PooledConnection.class.getName())
+                .toInstance(Level.OFF);
     }
 
     @Singleton